store_api/serializers.py
from rest_framework import serializers

from store_api.models import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'username', 'email']