requirements.txt
Django
djangorestframework
django-cors-headers
requests
PyJWT



